functions/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "earthquake-notifier"
version = "1.0.0"
dependencies = [
    "firebase-functions==0.4.2",
    "firebase-admin==6.2.0",
    "functions-framework==3.5.0",
    "requests==2.31.0",
    "pushbullet.py==0.12.0",
    "python-dotenv==1.0.0"
]
